Déplace `/api/nouvelleFonctionnalite/:id` dans les routes explicitement publiques
