+++ src/routes/routesApi.js
@@ -642,26 +642,6 @@ const routesApi = (
     }
   );
 
-  routes.get(
-    '/nouvelleFonctionnalite/:id',
-    middleware.aseptise('id'),
-    (requete, reponse) => {
-      const idNouvelleFonctionnalite = requete.params.id;
-      const nouvelleFonctionnalite = referentiel.nouvelleFonctionnalite(
-        idNouvelleFonctionnalite
-      );
-
-      if (!nouvelleFonctionnalite) {
-        reponse.status(404).send('Nouvelle fonctionnalité inconnue');
-        return;
-      }
-
-      reponse.render(
-        `nouvellesFonctionnalites/${nouvelleFonctionnalite.fichierPug}`
-      );
-    }
-  );
-
   return routes;
 };
